% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bi_chain_cpp}
\alias{bi_chain_cpp}
\title{bivariate MC chain}
\usage{
bi_chain_cpp(N, n = 100L, a = 50L, b = 50L)
}
\arguments{
\item{N}{sample size}

\item{n}{initial value}

\item{a}{sample size}

\item{b}{sample size}
}
\value{
mat A matrix
}
\description{
bivariate MC chain
}
