% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Title}
\usage{
generate.data(V, K, inter.p = 0.8, between.p = 0.2)
}
\arguments{
\item{between.p}{}
}
\value{
A \code{list} object comprising:
\item{X} Design matrix of predictors.
\item{Y} Response vector.
\item{Adj} Adjacency matrix.
}
\description{
Title
}
\examples{
V <- 100
K <- 3
dataset <- generate.data(V, K)
str(dataset)

}
