% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{bivariate MC chain}
\usage{
GibbsC(N, n = 100L, a = 50L, b = 50L)
}
\arguments{
\item{N}{sample size}

\item{n}{Parameter in the bivariate density.}

\item{a}{Parameter in the bivariate density.}

\item{b}{Parameter in the bivariate density.}
}
\value{
mat A matrix of random samples.
}
\description{
bivariate MC chain
}
