% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.net.R
\name{lm.net}
\alias{lm.net}
\title{Title}
\usage{
lm.net(X, Y, Adj, lambda, gamma)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{Adj}{}

\item{lambda}{}

\item{gamma}{}
}
\value{
A \code{list} object comprising:
\item{alpha} Vector of network cohesion.
\item{beta} Regression coefficients.
}
\description{
Title
}
\examples{

}
