% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq.net.R
\name{rq.net}
\alias{rq.net}
\title{Fitting Quantile Regression Models for Network-linked Data}
\usage{
rq.net(X, Y, Adj, lambda = 0.1, tau = 0.5)
}
\arguments{
\item{X}{Design matrix of predictors.}

\item{Y}{Response vector.}

\item{Adj}{Adjacency matrix.}

\item{lambda}{Regularization parameter.}

\item{tau}{The quantile to be estimated.}
}
\value{
A \code{list} object comprising:
\item{alpha}{Vector of network cohesion.} 
\item{beta}{Regression coefficients.}
}
\description{
Fitting Quantile Regression Models for Network-linked Data
}
\examples{
V <- 100
K <- 3
p <- 3
data <- generate.data(V, K, p)
rq.net(data$X, data$Y, data$Adj)
}
