\name{StatComp21060-package}
\alias{StatComp21060-package}
\docType{package}
\title{
 Regression models for network-linked data
}
\description{
This package aims to fit regression models for network-linked data
}
\author{
Borui Tang
Maintainer: Borui Tang <tangborui@mail.ustc.edu.cn>
}
\keyword{ package }