% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq.net.R
\name{rq.net}
\alias{rq.net}
\title{Title}
\usage{
rq.net(X, Y, Adj, lambda, tau = 0.5)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{Adj}{}

\item{lambda}{}

\item{tau}{}
}
\value{
A \code{list} object comprising:
\item{alpha} Vector of network cohesion.
\item{beta} Regression coefficients.
}
\description{
Title
}
\examples{


}
